Started work on new windows based design
