+++ WiSim/frontend/package.json
@@ -14,6 +14,7 @@
     "vite": "^6.3.2"
   },
   "dependencies": {
+    "@neodrag/svelte": "^2.3.2",
     "echarts": "^5.6.0"
   }
 }
