+++ WiSim/frontend/package.json.md5
@@ -1,1 +1,1 @@
-76132d18785fe85ba5775f1712068f88
+bb4113097f96c0e88797ac7cb51f0bd2
