+++ WiSim/frontend/src (old)/main.js
@@ -1,0 +1,7 @@
+import { mount } from 'svelte';
+import './style.css'
+import App from './App.svelte'
+
+const app = mount(App, { target: document.getElementById("app") });
+
+export default app
