+++ WiSim/frontend/src/style.css
@@ -6,7 +6,7 @@ html {
 
 body {
   margin: 0;
-  color: white;
+  color: black;
   font-family: "Nunito", -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto",
   "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
   sans-serif;
@@ -55,19 +55,18 @@ h2 {
 
 button {
   background-color: rgba(0, 0, 0, 0);
-  text-align: left;
-  color: rgba(255, 255, 255, 1);
+  text-align: center;
+  color: black;
   width: 120px;
-  height: 30px;
-  border-radius: 3px;
-  border: none;
+  padding: 5px 10px 5px 10px;
+  border: 3px solid;
   padding: 0 10px;
   cursor: pointer;
   font-size: 17px;
 }
 
 button:hover {
-  background-color: rgba(255, 255, 255, 0.3);
+  background-color: rgba(0, 0, 0, 0.5);
 }
 
 button:active {
@@ -98,4 +97,8 @@ button:active {
   overflow-y: scroll;
   height: calc(100vh - 20px - 50px);
   width: calc(100% - 240px)
+}
+
+.window_content {
+
 }
